Postgres initializer in separate module, correct parent pom
