+++ neighbour-model/src/test/java/no/vegvesen/ixn/federation/repository/NeighbourRepositorySelectorIT.java
@@ -1,9 +1,8 @@
 package no.vegvesen.ixn.federation.repository;
 
-import no.vegvesen.ixn.docker.PostgresTestcontainerInitializer;
+import no.vegvesen.ixn.postgresinit.PostgresTestcontainerInitializer;
 import no.vegvesen.ixn.federation.api.v1_0.SubscriptionStatus;
 import no.vegvesen.ixn.federation.model.*;
-import no.vegvesen.ixn.postgresinit.PostgresTestcontainerInitializer;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
