+++ neighbour-service/src/test/java/no/vegvesen/ixn/federation/service/NeighbourDiscovererIT.java
@@ -1,7 +1,7 @@
 package no.vegvesen.ixn.federation.service;
 
 
-import no.vegvesen.ixn.docker.PostgresTestcontainerInitializer;
+import no.vegvesen.ixn.federation.api.v1_0.CapabilityApi;
 import no.vegvesen.ixn.federation.api.v1_0.Datex2DataTypeApi;
 import no.vegvesen.ixn.federation.api.v1_0.SubscriptionStatus;
 import no.vegvesen.ixn.federation.discoverer.DNSFacade;
@@ -9,7 +9,7 @@ import no.vegvesen.ixn.federation.discoverer.facade.NeighbourRESTFacade;
 import no.vegvesen.ixn.federation.model.*;
 import no.vegvesen.ixn.federation.repository.NeighbourRepository;
 import no.vegvesen.ixn.federation.repository.SelfRepository;
-import no.vegvesen.ixn.federation.service.NeighbourService;
+import no.vegvesen.ixn.postgresinit.PostgresTestcontainerInitializer;
 import no.vegvesen.ixn.properties.MessageProperty;
 import org.assertj.core.util.Lists;
 import org.assertj.core.util.Sets;
