+++ testcontainers-init-postgres/pom.xml
@@ -7,7 +7,7 @@
     <parent>
         <groupId>net.nordicway.interchange</groupId>
         <artifactId>interchange-parent</artifactId>
-        <version>1.0-SNAPSHOT</version>
+        <version>1.0.3-SNAPSHOT</version>
     </parent>
     <artifactId>testcontainers-init-postgres</artifactId>
